backend/myrequirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioread==3.1.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
edge-tts==7.2.7
faiss-cpu==1.13.1
fastapi==0.109.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.45.0
google-generativeai==0.4.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.62.3
gTTS==2.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.4
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.53
langsmith==0.1.147
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==2.1.5
marshmallow==3.26.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==1.26.4
openai==1.12.0
opencv-python==4.9.0.80
orjson==3.11.5
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
propcache==0.4.1
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pypdfium2==5.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
PyYAML==6.0.3
regex==2025.11.3
requests==2.28.1
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.12.0
sentence-transformers==2.5.1
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.45
starlette==0.36.3
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.2.0+cpu
torchvision==0.17.0+cpu
tqdm==4.67.1
transformers==4.38.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.13
uvicorn==0.27.1
watchfiles==1.1.1
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.22.0
